// src/types/employee/salary/debit/list.tsx
import DebitListStatus from "../../../../enums/employee/salary/debit/list";

export interface Debit {
  id: number;
  personel_id: number;
  aylik_ucret: string;
  odeme_sekli: string;
  maas_sayisi: number;
  baslangic_tarihi: string;
  olusturulma_tarihi: string;
  guncellenme_tarihi: string;
  platform_id: number;
  updated_at: string;
  created_At: string;
  maas_ayi: string;
}

export interface ILink {
  url: string | null;
  label: string;
  active: boolean;
}

export interface DebitListPaginate {
  current_page: number;
  first_page_url: string;
  from: number | null;
  last_page: number;
  last_page_url: string;
  next_page_url: string | null;
  links: Debit[];
  path: string;
  per_page: number;
  prev_page_url: string | null;
  to: number | null;
  total: number;
}

export interface DebitListResponse {
  data: Debit[];
  current_page: number;
  first_page_url: string;
  from: number | null;
  last_page: number;
  last_page_url: string;
  links: ILink[];
  next_page_url: string | null;
  path: string;
  per_page: number;
  prev_page_url: string | null;
  to: number | null;
  total: number;
}

export interface DebitState {
  data: Debit[] | null;
  status: DebitListStatus;
  error: string | null;
}

export interface DebitListArgs {
  enabled?: boolean;
  [key: string]: any;
}